Fix fps limiter running at the start of each frame
This means that the limiter uses the render time of the *last* frame to determine how long it should wait for the current frame.
This is especially bad when the last frame had no client tick but the current frame has a client tick. Resulting in the current frame being displayed too late and the next frame being displayed too early.